tests/wpt/tests/workers/constructors/SharedWorker/null-arguments.html

<!doctype html>
<title>null as arguments</title>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script>
async_test(t => {
  const worker = new SharedWorker(null, null);
  worker.port.onmessage = t.step_func_done(e => {
    assert_equals(e.data[0], 'null', 'first arg (script name)');
    assert_equals(e.data[1], '', 'second arg (worker name)');
  });
}, 'Test constructing a shared worker with null');
</script>
